16-youtube-summarizer/requirements.txt

fastapi==0.116.1
starlette==0.47.1
uvicorn==0.35.0
python-dotenv==1.1.1
google-genai==1.25.0
youtube-transcript-api==1.1.1
pytest==8.4.1
httpx==0.28.1
pytest-asyncio==1.0.0
